refactor: update scale down model instance pick strategy

+++ gpustack/server/controllers.py
@@ -1,7 +1,14 @@
+import copy
 import logging
 import random
 import string
+from typing import List
 from sqlmodel.ext.asyncio.session import AsyncSession
+from gpustack.config.config import Config
+from gpustack.policies.offload_layer_policy import OffloadLayerPolicy
+from gpustack.policies.placement_policy import PlacementPolicy, ScaleTypeEnum
+from gpustack.policies.policy import ModelInstanceScore
+from gpustack.policies.status_policy import StatusPolicy
 from gpustack.schemas.models import (
     Model,
     ModelInstance,
@@ -17,8 +24,10 @@ logger = logging.getLogger(__name__)
 
 
 class ModelController:
-    def __init__(self):
+    def __init__(self, cfg: Config):
         self._engine = get_engine()
+        self._config = cfg
+
         pass
 
     async def start(self):
@@ -37,14 +46,16 @@ class ModelController:
         model: Model = event.data
         try:
             async with AsyncSession(self._engine) as session:
-                await sync_replicas(session, model)
+                await sync_replicas(session, model, self._config)
         except Exception as e:
             logger.error(f"Failed to reconcile model {model.name}: {e}")
 
 
 class ModelInstanceController:
-    def __init__(self):
+    def __init__(self, cfg: Config):
         self._engine = get_engine()
+        self._config = cfg
+
         pass
 
     async def start(self):
@@ -68,7 +79,7 @@ class ModelInstanceController:
                     return
 
                 if event.type == EventType.DELETED:
-                    await sync_replicas(session, model)
+                    await sync_replicas(session, model, self._config)
 
                 await model.refresh(session)
                 await sync_ready_replicas(session, model)
@@ -79,7 +90,7 @@ class ModelInstanceController:
             )
 
 
-async def sync_replicas(session: AsyncSession, model: Model):
+async def sync_replicas(session: AsyncSession, model: Model, cfg: Config):
     """
     Synchronize the replicas.
     """
@@ -107,9 +118,60 @@ async def sync_replicas(session: AsyncSession, model: Model):
             logger.debug(f"Created model instance for model {model.name}")
 
     elif len(instances) > model.replicas:
-        for instance in instances[model.replicas :]:
-            await instance.delete(session)
-            logger.debug(f"Deleted model instance {instance.name}")
+        candidates = await find_scale_down_candidates(instances, model)
+
+        scale_down_count = len(candidates) - model.replicas
+        if scale_down_count > 0:
+            for candidate in candidates[:scale_down_count]:
+                instance = candidate.instance
+                await instance.delete(session)
+                logger.debug(f"Deleted model instance {instance.name}")
+
+
+async def policy_score_instances(
+    policy, instances: List[ModelInstance]
+) -> List[ModelInstanceScore]:
+    return await policy.score_instances(copy.deepcopy(instances))
+
+
+async def find_scale_down_candidates(
+    instances: List[ModelInstance], model: Model
+) -> List[ModelInstanceScore]:
+    try:
+        placement_policy = PlacementPolicy(model, scale_type=ScaleTypeEnum.SCALE_DOWN)
+        placement_candidates = await policy_score_instances(placement_policy, instances)
+
+        offload_layer_policy = OffloadLayerPolicy(model)
+        offload_candidates = await policy_score_instances(
+            offload_layer_policy, instances
+        )
+
+        status_policy = StatusPolicy(model)
+        status_candidates = await policy_score_instances(status_policy, instances)
+
+        offload_cand_map = {cand.model_instance.id: cand for cand in offload_candidates}
+        placement_cand_map = {
+            cand.model_instance.id: cand for cand in placement_candidates
+        }
+
+        for cand in status_candidates:
+            score = cand.score * 100
+            offload_candidate = offload_cand_map.get(cand.model_instance.id)
+            score += offload_candidate.score * 10 if offload_candidate else 0
+
+            placement_candidate = placement_cand_map.get(cand.model_instance.id)
+            score += placement_candidate.score if placement_candidate else 0
+            cand.score = score / 111
+
+        final_candidates = sorted(
+            status_candidates, key=lambda x: x.score, reverse=False
+        )
+        return final_candidates
+    except Exception as e:
+        state_message = (
+            f"Failed to find scale down candidates for model {model.name}: {e}"
+        )
+        logger.error(state_message)
 
 
 async def sync_ready_replicas(session: AsyncSession, model: Model):
